[compat] add legacy schema dsl behavior
